fix declaratio/definition/extern type mismatch, finally MSVC proved to help in solving those, for some reason nor clang neither gcc complained about them

+++ code/logistics.cpp
@@ -444,8 +444,8 @@ void Logistics::render (void)
 
 //-----------------------------------------------------------------------------
 
-extern unsigned long MultiPlayTeamId;
-extern unsigned long MultiPlayCommanderId;
+extern unsigned int MultiPlayTeamId;
+extern unsigned int MultiPlayCommanderId;
 
 int _stdcall Logistics::beginMission(void*, int, void*[])
 {
